+++ .eslintrc.json
@@ -28,6 +28,5 @@
     "node/no-unsupported-features/es-syntax": "off",
     "no-underscore-dangle": "off",
     "import/prefer-default-export": "off"
-  },
-
+  }
 }
