+++ src/resources/schedules/schedule.model.js
