+++ src/resources/schedules/schedule.service.js
